# openapi.yaml
openapi: 3.1.0
info:
  title: Домовой — API приложения дома в MAX
  version: 1.0.0
  description: 'API мини-приложения и бота MAX «Домовой»: вход по QR-коду квитанции, обращения в УК, начисления,
    счётчики, лента и опросы дома, раздел председателя, кабинет УК, подбор дома.


    **Как войти для проверки.** Без квитанции и без MAX — ролью демо-дома: `POST /api/demo/roles/{key}/take` (key:
    chairman, owner12, member12, owner45, owner78, newcomer30) вернёт `token`; дальше заголовок `Authorization:
    Bearer <token>`. Кабинет УК — `POST /api/dispatcher/login` с логином демо-УК (пароль — на служебном слайде презентации),
    тоже `token`.


    **Ошибки** — JSON `{ "error": "код", "message": "текст для человека" }`.


    **Модель доступа.** Квитанция — заявка, а не пропуск: новая квартира приходит со статусом `pending`. Уровень
    0 (pending) — свои начисления, счётчики и обращения в УК сразу. Уровень 1 (active) — соседи, лента, опросы.
    Уровень 2 — дом целиком, только председателю. Обращение удалить нельзя.


    Кабинет оператора (`/api/admin/*`) — внутренний и здесь не описан.'
servers:
- url: https://domovoy.devcore.com.ru
  description: Бой
- url: http://localhost:3000
  description: Локальный запуск (docker compose)
...
security:
- bearerAuth: []
- cookieAuth: []
paths:
...
  /api/feed:
    get:
      tags:
      - Дом
      summary: Лента дома
      description: Объявления УК и председателя, отключения, доска соседей. Уровень доступа 1 — после подтверждения.
      operationId: get_api_feed
      parameters:
      - name: category
        in: query
        required: false
        schema:
          type: string
      - name: scope
        in: query
        required: false
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Успех
          content:
            application/json:
              schema:
                type: object
                properties:
                  posts:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        type:
                          type:
                          - string
                          - 'null'
                        category:
                          type:
                          - string
                          - 'null'
                        title:
                          type:
                          - string
                          - 'null'
                        body:
                          type:
                          - string
                          - 'null'
                        contact:
                          type:
                          - string
                          - 'null'
                        publishedAt:
                          type:
                          - string
                          - 'null'
                          format: date-time
                        expiresAt:
                          type:
                          - string
                          - 'null'
                          format: date-time
                        authorName:
                          type:
                          - string
                          - 'null'
                        authorId:
                          type:
                          - string
                          - 'null'
                        chairmanName:
                          type:
                          - string
                          - 'null'
                        phone:
                          type:
                          - string
                          - 'null'
                        maxUsername:
                          type:
                          - string
                          - 'null'
                        mine:
                          type:
                          - boolean
                          - 'null'
                        categoryLabel:
                          type:
                          - string
                          - 'null'
                        hasPhoto:
                          type:
                          - boolean
                          - 'null'
                        unread:
                          type:
                          - boolean
                          - 'null'
                        expired:
                          type:
                          - boolean
                          - 'null'
                        author:
                          type:
                          - string
                          - 'null'
                      required:
                      - id
                      - type
                      - category
                      - title
                      - body
                      - contact
                      - publishedAt
                      - expiresAt
                      - authorName
                      - authorId
                      - chairmanName
                      - phone
                      - maxUsername
                      - mine
                      - categoryLabel
                      - hasPhoto
                      - unread
                      - expired
                      - author
                  total:
                    type: number
                required:
                - posts
                - total
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
      - Дом
      summary: Объявление соседям
      description: Житель публикует объявление на доске «Соседи предлагают».
      operationId: post_api_feed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                propertyId:
                  type:
                  - string
                  - 'null'
                title:
                  type:
                  - string
                  - 'null'
                body:
                  type:
                  - string
                  - 'null'
                contact:
                  type:
                  - string
                  - 'null'
                sharePhone:
                  type:
                  - boolean
                  - 'null'
                category:
                  type:
                  - string
                  - 'null'
              required:
              - propertyId
              - title
              - body
      responses:
        '201':
          description: Создано
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  id:
                    type: string
                required:
                - status
                - id
        '400':
          description: Неверные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/house/claim:
    post:
      tags:
      - Житель
      summary: Подключить дом без УК
      description: 'Заявка оператору: у ТСЖ, ЖСК и домов без лицензии некому назначить председателя.'
      operationId: post_api_house_claim
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                propertyId:
                  type:
                  - string
                  - 'null'
                note:
                  type:
                  - string
                  - 'null'
              required:
              - propertyId
      responses:
        '200':
          description: Успех
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type:
                    - boolean
                    - 'null'
                  created:
                    type:
                    - boolean
                    - 'null'
                required:
                - ok
                - created
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/polls:
    get:
      tags:
      - Дом
      summary: Опросы дома
      description: Юридической силы общего собрания не имеют.
      operationId: get_api_polls
      responses:
        '200':
          description: Успех
          content:
            application/json:
              schema:
                type: object
                properties:
                  polls:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        title:
                          type:
                          - string
                          - 'null'
                        description: {}
                        closesAt:
                          type:
                          - string
                          - 'null'
                          format: date-time
                        closed:
                          type:
                          - boolean
                          - 'null'
                        myOptionId: {}
                        total:
                          type: number
                        showResults:
                          type:
                          - boolean
                          - 'null'
                        legalNotice:
                          type:
                          - string
                          - 'null'
                        options:
                          type: array
                          items:
                            type: object
                            properties:
                              id:
                                type: string
                              text:
                                type:
                                - string
                                - 'null'
                              votes: {}
                              percent: {}
                            required:
                            - id
                            - text
                            - votes
                            - percent
                      required:
                      - id
                      - title
                      - description
                      - closesAt
                      - closed
                      - myOptionId
                      - total
                      - showResults
                      - legalNotice
                      - options
                required:
                - polls
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/polls/{id}:
    get:
      tags:
      - Дом
      summary: Опрос
      description: Варианты и результаты (после своего голоса или по завершении).
      operationId: get_api_polls_id
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Успех
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  title:
                    type:
                    - string
                    - 'null'
                  description: {}
                  closesAt:
                    type:
                    - string
                    - 'null'
                    format: date-time
                  closed:
                    type:
                    - boolean
                    - 'null'
                  myOptionId: {}
                  total:
                    type: number
                  showResults:
                    type:
                    - boolean
                    - 'null'
                  legalNotice:
                    type:
                    - string
                    - 'null'
                  options:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        text:
                          type:
                          - string
                          - 'null'
                        votes: {}
                        percent: {}
                      required:
                      - id
                      - text
                      - votes
                      - percent
                required:
                - id
                - title
                - description
                - closesAt
                - closed
                - myOptionId
                - total
                - showResults
                - legalNotice
                - options
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/polls/{id}/vote:
    post:
      tags:
      - Дом
      summary: Проголосовать
      description: Голос можно менять, пока опрос идёт.
      operationId: post_api_polls_id_vote
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                optionId:
                  type:
                  - string
                  - 'null'
              required:
              - optionId
      responses:
        '200':
          description: Успех
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  title:
                    type:
                    - string
                    - 'null'
                  description: {}
                  closesAt:
                    type:
                    - string
                    - 'null'
                    format: date-time
                  closed:
                    type:
                    - boolean
                    - 'null'
                  myOptionId:
                    type:
                    - string
                    - 'null'
                  total:
                    type: number
                  showResults:
                    type:
                    - boolean
                    - 'null'
                  legalNotice:
                    type:
                    - string
                    - 'null'
                  options:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        text:
                          type:
                          - string
                          - 'null'
                        votes:
                          type:
                          - number
                          - 'null'
                        percent:
                          type:
                          - number
                          - 'null'
                      required:
                      - id
                      - text
                      - votes
                      - percent
                required:
                - id
                - title
                - description
                - closesAt
                - closed
                - myOptionId
                - total
                - showResults
                - legalNotice
                - options
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/posts/{id}/photo:
    delete:
      tags:
      - Дом
      summary: Убрать фото объявления
      operationId: delete_api_posts_id_photo
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
              - file
      responses:
        '200':
          description: Успех
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    const: ok
                required:
                - status
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Нет прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      tags:
      - Дом
      summary: Фото объявления
      operationId: get_api_posts_id_photo
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
              - file
      responses:
        '200':
          description: Успех
          content:
            image/png:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Нет прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
      - Дом
      summary: Загрузить фото объявления
      description: Одно изображение до 10 МБ.
      operationId: post_api_posts_id_photo
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
              - file
      responses:
        '200':
          description: Успех
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                required:
                - status
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Нет прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/posts/{id}/read:
    post:
      tags:
      - Дом
      summary: Отметить прочитанным
      operationId: post_api_posts_id_read
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Успех
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                required:
                - status
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/properties/{propertyId}/house-contacts:
    get:
      tags:
      - Житель
      summary: Телефоны дома
      description: Аварийные и служебные телефоны дома, добавленные УК или председателем.
      operationId: get_api_properties_propertyId_house_contacts
      parameters:
      - name: propertyId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Успех
          content:
            application/json:
              schema:
                type: object
                properties:
                  contacts:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        kind:
                          type:
                          - string
                          - 'null'
                        title:
                          type:
                          - string
                          - 'null'
                        phone:
                          type:
                          - string
                          - 'null'
                        note:
                          type:
                          - string
                          - 'null'
                        updatedByRole:
                          type:
                          - string
                          - 'null'
                        updatedAt:
                          type:
                          - string
                          - 'null'
                          format: date-time
                      required:
                      - id
                      - kind
                      - title
                      - phone
                      - note
                      - updatedByRole
                      - updatedAt
                required:
                - contacts
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Нет прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Токен сессии из /api/auth/qr, /api/auth/max, /api/demo/roles/{key}/take или /api/dispatcher/login
    cookieAuth:
      type: apiKey
      in: cookie
      name: zd_session
      description: Та же сессия в cookie — её ставит сервер при входе
...
  schemas:
    Error:
      type: object
      required:
      - error
      properties:
        error:
          type: string
          description: Машинный код ошибки
        message:
          type: string
          description: Понятный человеку текст
  responses:
    Unauthorized:
      description: Нужен вход
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'